new option to either autodelete or not after automatic processing

+++ bin/batchcachedir_check_frame.sh
@@ -1,9 +1,12 @@
 #!/bin/bash
 PROC=0
-if [ -z $1 ]; then echo "set parameter - frame"; echo "if second parameter is 1, this script will perform reprocessing automatically"; exit; fi
+AUTODEL=1
+if [ -z $1 ]; then echo "set parameter - frame"; echo "if second parameter is 1, this script will perform reprocessing automatically";
+ echo "third parameter is to set autodelete in case all is checked OK - it is set to 1 by default, so careful"; exit; fi
 if [ ! -d $1 ]; then echo "you need to be in a folder (e.g. your BATCH_CACHE_DIR) with this frame data"; exit; fi
 frame=$1
 if [ ! -z $2 ]; then PROC=$2; fi
+if [ ! -z $3 ]; then AUTODEL=$2; fi
 #cd $BATCH_CACHE_DIR
 todel=0
 
@@ -120,4 +123,6 @@ else
    fi
 fi
 
+if [ $AUTODEL == 1 ]; then
 if [ $todel == 1 ]; then echo "this frame dir will be deleted now: " $frame; rm -rf $frame; fi
+fi
